chirag/requirements.txt
Flask
Pillow
python-dotenv
requests